Function astar_collator::parachain::service::new_partial

pub fn new_partial<RuntimeApi, Executor, BIQ>(
    config: &Configuration, 
    build_import_queue: BIQ
) -> Result<PartialComponents<TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>, TFullBackend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, (), DefaultImportQueue<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, FullPool<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, (Option<Telemetry>, Option<TelemetryWorkerHandle>, Arc<Backend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>>)>, Error> where
    RuntimeApi: ConstructRuntimeApi<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: TaggedTransactionQueue<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>> + Metadata<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>> + SessionKeys<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>> + ApiExt<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, StateBackend = StateBackendFor<TFullBackend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>> + OffchainWorkerApi<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>> + BlockBuilder<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>> + EthereumRuntimeRPCApi<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>,
    StateBackendFor<TFullBackend<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>, Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>>: StateBackend<BlakeTwo256>,
    Executor: NativeExecutionDispatch + 'static,
    BIQ: FnOnce(Arc<TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, FrontierBlockImport<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, Arc<TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, &Configuration, Option<TelemetryHandle>, &TaskManager) -> Result<DefaultImportQueue<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, TFullClient<Block<Header<u32, BlakeTwo256>, OpaqueExtrinsic>, RuntimeApi, NativeElseWasmExecutor<Executor>>>, Error>, 

Starts a ServiceBuilder for a full service.

Use this macro if you don’t actually need the full service, but just the builder in order to be able to perform chain operations.